+++ internal/connectors/proxmox/proxmox.go
@@ -73,6 +73,7 @@ func (c *Connector) Capabilities() []connector.Capability {
 		connector.CapabilityDestroy,
 		connector.CapabilityTemplateBuild,
 		connector.CapabilityNodePrerequisites,
+		connector.CapabilityGuestAgent,
 	}
 }